+++ agrest-base/src/main/java/io/agrest/protocol/Dir.java
@@ -6,5 +6,17 @@ package io.agrest.protocol;
  * @since 2.13
  */
 public enum Dir {
-    ASC, ASC_CI, DESC, DESC_CI;
+    ASC,
+
+    /**
+     * @since 3.4
+     */
+    ASC_CI,
+
+    DESC,
+
+    /**
+     * @since 3.4
+     */
+    DESC_CI;
 }