+++ agrest-client/src/main/java/io/agrest/client/protocol/Sort.java
@@ -1,43 +1,47 @@
 package io.agrest.client.protocol;
 
+import io.agrest.protocol.Dir;
+
 /**
  * @since 2.0
  */
 public class Sort {
 
-	public enum SortDirection {
-		ASCENDING("ASC"), DESCENDING("DESC");
-
-		private String abbrev;
-
-		SortDirection(String abbrev) {
-			this.abbrev = abbrev;
-		}
-
-		public String abbrev() {
-			return abbrev;
-		}
-	}
-
 	public static Sort property(String propertyName) {
-		return new Sort(propertyName, SortDirection.ASCENDING);
+		return new Sort(propertyName, Dir.ASC);
 	}
 
 	private String propertyName;
-	private SortDirection direction;
+	private Dir direction;
 
-	private Sort(String propertyName, SortDirection direction) {
+	private Sort(String propertyName, Dir direction) {
 		this.propertyName = propertyName;
 		this.direction = direction;
 	}
 
 	public Sort asc() {
-		direction = SortDirection.ASCENDING;
+		direction = Dir.ASC;
 		return this;
 	}
 
 	public Sort desc() {
-		direction = SortDirection.DESCENDING;
+		direction = Dir.DESC;
+		return this;
+	}
+
+    /**
+     * @since 3.4
+     */
+	public Sort ascCi() {
+		direction = Dir.ASC_CI;
+		return this;
+	}
+
+    /**
+     * @since 3.4
+     */
+	public Sort descCi() {
+		direction = Dir.DESC_CI;
 		return this;
 	}
 
@@ -45,7 +49,7 @@ public class Sort {
 		return propertyName;
 	}
 
-	public SortDirection getDirection() {
+	public Dir getDirection() {
 		return direction;
 	}
 }